fix(help): Shorten --help help to shorten help
Applications may want to constrain their short help to use the two-column
format so the -h output is as compact as possible.  They might do this by
optimizing short help texts to fit within the right-hand column on
80-character terminals, and then providing long help texts with more info.
However, c1c269b427 made the short help text for -h/--help much longer
when the short and long help are different, forcing short help to wrap to
the multiple-line format if medium-sized option/value names are used.

Reword the short help for this case to be as compact as possible.  The
net length is still a few characters longer than in clap 3, but this still
substantially increases the maximum width of the left column while
continuing to document the difference between -h and --help.

Fixes #4409