test: Add failing test for issue #4520
This test is semantically equivalent to the situation described in that
issue's description:
a first option conflicts with a second option.
a third option requires the second option.
Clap lets us provide the first and the third options only, which
shouldn't be allowed.

The test fails here, but is fixed by the next commit in this PR.