refactor(help): Simplify a combined next_help_line/use_long condition
The thing this logic really cares about (as the comment says) is
that we're printing on the next line (`next_line_help`), not that
the help output overall is in long form (`use_long`) as opposed to
short form.  So for this code to cleanly say what it means, it
should just refer to `next_line_help`, not also to `use_long`.

And in fact the cleaner version has exactly the same behavior as the
existing version: whenever `use_long` is true, `next_line_help` here
will also be true.

That's because this function's caller's caller gets `next_line_help`
from `arg_next_line_help`, and that returns true whenever `use_long`
is true.

So, simplify this condition to the cleaner, equivalent, form.