test(help): Test a hidden option is hidden in template, too
When first reading the relevant code in src/output/help_template.rs,
I thought there was a bug where options that should be hidden
would get shown after all when a template was used.

As this test shows, there is not such a bug.  But that fact is
a bit fragile: while the `write_args` call sites in `write_all_args`
all filter the args with `should_show_arg`, the call sites in
`write_templated_help` don't.  So, add this line to the tests
to help prevent such a bug sneaking in in some future refactor.