refactor(help): Simplify write_args, filtering for should-show up front
It's already the case that in each step of this function, any arg that
it does anything with is one that has already been filtered by
should_show_arg.  But seeing that is a bit tangly.

This hopefully makes that fact significantly more transparent.
It also replaces the name `ord_v` which seems hard to understand;
and as a bonus, the code gets shorter.