test: Show `conflicts_with` kebab-case failure
Add a failing test demonstrating that `conflicts_with` fails when given
a kebab-case argument.

Given that multi-word arguments are referred to using kebab-case for
their long names and when parsed from the command line, it is reasonable
to expect that they can be referred to using kebab case when specified
in a `conflicts_with` argument.

This worked in Clap v3 (or at least v3.1.8, see
https://github.com/stellar/soroban-tools/pull/491)