fix: Remove strict linting from generated code
The `#[deny(correctness)]` attribute is inside some generated code (such
as the `Subcommand` derive macro). As of Rust 1.69.0, this is triggering
a clippy error in some uses. Due to the placement of the `deny`, it is
not possible for a user of clap to override it.

This change removes all the instances of this `deny` attribute. The same
was done on the `master` branch in #4739.

Fixes: #4857