fix(derive): Allow `clippy::almost_swapped`
As of Rust 1.69, the `clippy::almost_swapped` lint was enhanced and now
triggers on certain derive macros inside clap. To avoid the need for the
user to allow these lints outside of the macro (which generally requires
a module level `allow`), the allow was added to the generated code for
the v4 branch in #4735.

This change backports that fix the the `v3-master` branch.

Fixes: #4857