fix(parser): Don't include groups in argument suggestions
If we are assuming the arguments are present, then there is no reason to
show alternatives from the group.

There may be other ways large groups show up.
For now, I'm taking the whack-a-mole approach of handling them on a
case-by-case basis.
When we get to one where the group is relevant to show,
we should handle a general case then.

Fixes #5958