docs(complete): Fix sourcing completions for pwsh
The previous instructions were wrong in that they set the `COMPLETE`
env var only while the `$PROFILE` script was written. This was akin to

    COMPLETE=bash echo "source <(your_program)" >> ~/.bashrc

whereas the corrected version sets `COMPLETE` as part of the `$PROFILE`
script, mirroring the instructions for sourcing completion in the other
shells.

Extending `$PROFILE` in this way also requires that scripts are allowed
to be executed. This requires changing the execution policy from its
default value `Restricted` on Windows. There's no such issue when using
PowerShell on macOS or Linux.