clap_complete: add unit test for fish env completer
Related to: https://github.com/clap-rs/clap/issues/6196

This PR adds a unit test showing the current behavior of the the Fish
shell env completer, that quotes paths incorrectly when there are spaces
in the path. The issue should be demonstrated in this PR, and a
follow-up PR could have a fix for the issue.